Users delete

+++ app/Http/Livewire/Users/UsersTable.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Users;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\View\View;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -11,6 +12,8 @@ class UsersTable extends Component
 {
     use WithPagination;
 
+    public $deleted = false;
+    
     protected $listeners = ['delete'];
 
     public function render(): View
@@ -20,8 +23,12 @@ class UsersTable extends Component
         return view('livewire.users.users-table', compact('users'))->layout('layout.app');
     }
 
-    public function delete(User $user): void
+    public function delete(User $user, $pass): void
     {
-        $user->delete();
+        if ($pass == Auth::user()->password) {
+            $user->delete();
+
+            $this->deleted = true;
+        }
     }
 }
